Rename cmd to src
I want one dir so the eventual tests can be ran by just pointing
to one directory, but I think the way I was using cmd was
unconventional and same would go for pkg.
