+++ scripts/build.sh
@@ -3,5 +3,5 @@ set -eu
 
 ROOT=$(pwd)
 mkdir -p "$ROOT/bin"
-go build -C cmd -o "$ROOT/bin/$APP_NAME"
+go build -C src -o "$ROOT/bin/$APP_NAME"
 chmod +x "$ROOT/bin/$APP_NAME"
